requirements.txt
akshare>=1.11.0
yfinance>=0.2.0
pandas>=2.0.0
streamlit>=1.28.0
scikit-learn>=1.3.0
feedparser>=6.0
beautifulsoup4>=4.12.0
requests>=2.31.0
apscheduler>=3.10.0
jieba>=0.42.0
pyyaml>=6.0
numpy>=1.24.0